+++ src/controllers/contacts/addContact.js
@@ -1,4 +1,4 @@
-const serviceContacts = require("../../repository/contacts");
+const serviceContacts = require("../../services/contacts");
 
 const addContact = async (req, res) => {
   const { id: owner } = req.user;
