+++ src/controllers/contacts/deleteContactById.js
@@ -1,5 +1,5 @@
 const { createError } = require("../../helpers");
-const serviceContacts = require("../../repository/contacts");
+const serviceContacts = require("../../services/contacts");
 
 const deleteContactById = async (req, res) => {
   const { id } = req.params;
