+++ src/controllers/contacts/updateContactsById.js
@@ -1,5 +1,5 @@
 const { createError } = require("../../helpers");
-const serviceContacts = require("../../repository/contacts");
+const serviceContacts = require("../../services/contacts");
 
 const updateContactById = async (req, res) => {
   const { id } = req.params;
