+++ src/middlewares/validationToken.js
@@ -1,5 +1,5 @@
 const jwt = require("jsonwebtoken");
-const serviceUsers = require("../repository/users");
+const serviceUsers = require("../services/users");
 const { createError } = require("../helpers");
 const { SECRET_KEY } = process.env;
