+++ src/services/users/updateUserAvatarById.js
